committable

Trait CommitmentBoundsArkless

pub trait CommitmentBoundsArkless:
    AsRef<[u8]>
    + Clone
    + Copy
    + Debug
    + Eq
    + Hash
    + PartialEq
    + Send
    + Sync
    + 'static {
    // Required method
    fn default_commitment_no_preimage() -> Self;
}

Consolidate trait bounds for cryptographic commitments.

Downstream users should prefer CommitmentBounds, which is defined appropriately according to feature flags “ark-serialize”, “serde”.

“serde” is additive with “ark-serialize”, so if “serde” is enabled then so is “ark-serialize”. But it is possible to have “ark-serialize” without “serde”.

Required Methods

fn default_commitment_no_preimage() -> Self

Create a default commitment with no preimage.

Alternative to Default

Commitment does not impl Default so as to prevent users from accidentally creating a commitment that has no preimage. Sometimes, however, such a commitment is needed so we provide this convenience method. Even without this method, we cannot stop users from creating such a commitment using [CanonicalDeserialize].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.